refac: feature단위 폴더구조 폐지, 파일 용도 별 디렉토리 구조로 편입
