+++ src/components/register/manager/type.d.ts
@@ -1,4 +1,4 @@
-import { CenterNameResponse } from '@/features/common/types/center';
+import { CenterNameResponse } from '@/types/common/center';
 import {
   FieldError,
   FieldValues,
