--- src/features/common/types/index.d.ts
@@ -1,7 +1,0 @@
-export interface UpLoadFileRequest {
-  file: File[] | File;
-}
-
-export interface UpLoadFileResponse {
-  file_url: string;
-}
