+++ src/types/common/center.d.ts
