+++ src/types/common/index.d.ts
@@ -22,3 +22,11 @@ export interface ServerBusinessError {
   timeStamp: string;
   violations?: string[];
 }
+
+export interface UpLoadFileRequest {
+  file: File[] | File;
+}
+
+export interface UpLoadFileResponse {
+  file_url: string;
+}
