+++ src/types/common/lector.d.ts
